refactor(aad): adjust the pop up window for aad update manifest command
